fixed Customer, CustomerAddress, Order and OrderController...still work in progress to complete orders detail/edit/new responses.

+++ CannDash.API/App_Start/Controllers/OrdersController.cs
@@ -84,28 +84,30 @@ namespace CannDash.API.Controllers
                 return BadRequest();
             }
 
-            db.Entry(order).State = EntityState.Modified;
-
-            // delete the product orders in the existing order which no longer
-            // exist in the given order
             var dbOrder = db.Orders.Find(id);
-            foreach (var productOrder in dbOrder.ProductOrders)
-                if (order.ProductOrders
-                    .All(p => p.ProductOrderId != productOrder.ProductOrderId))
-                    db.ProductOrders.Remove(productOrder);
-
-            // update or insert the product orders from the given order, in the existing order
-            foreach (var productOrder in order.ProductOrders)
+            if (order.ProductOrders != null)
             {
-                var existing =
-                    dbOrder.ProductOrders.FirstOrDefault(
-                        p => p.ProductOrderId == productOrder.ProductOrderId);
-                System.Diagnostics.Trace.WriteLine("***productOrder " + productOrder.ProductOrderId + ": " + existing);
-                if (existing != null)
-                    db.Entry(existing).CurrentValues.SetValues(productOrder);
-                else
-                    dbOrder.ProductOrders.Add(productOrder);
+                // delete the product orders in the existing order which no longer
+                // exist in the given order
+                foreach (var productOrder in dbOrder.ProductOrders.ToList())
+                    if (order.ProductOrders
+                        .All(p => p.ProductOrderId != productOrder.ProductOrderId))
+                        db.ProductOrders.Remove(productOrder);
+
+                // update or insert the product orders from the given order, in the existing order
+                foreach (var productOrder in order.ProductOrders)
+                {
+                    var existing =
+                        dbOrder.ProductOrders.FirstOrDefault(
+                            p => p.ProductOrderId == productOrder.ProductOrderId);
+                    if (existing != null)
+                        db.Entry(existing).CurrentValues.SetValues(productOrder);
+                    else
+                        dbOrder.ProductOrders.Add(productOrder);
+                }
             }
+
+            db.Entry(dbOrder).CurrentValues.SetValues(order);
 
             try
             {
